third_party/WebKit/LayoutTests/http/tests/performance-timing/element-timing/basetest.html

<!DOCTYPE html>
<head>
<title>Performance Element Timing Test: Existence of Text Element Timing Entries</title>
</head>
<body>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>

    <h1 elementtiming="element1">Element 1</h1>
    <div elementtiming="element2">Element 2</div>
    <p elementtiming="element3">Element 3</p>
    <p>No timing</p>

    <script>
      async_test(function (t) {
        testRunner.capturePixelsAsyncThen(t.step_func_done(function() {
            var bufferedEntries = performance.getEntriesByType('elementtiming');
            assert_equals(bufferedEntries.length, 3, "Expected 3 elementtiming entries");
            bufferedEntries.sort(function(a, b) { return a.name < b.name ? -1 : (a.name > b.name ? 1 : 0); })

            assert_equals(bufferedEntries[0].entryType, "elementtiming");
            assert_equals(bufferedEntries[0].name, "element1");
            assert_equals(bufferedEntries[0].duration, 0);
            assert_equals(bufferedEntries[1].entryType, "elementtiming");
            assert_equals(bufferedEntries[1].name, "element2");
            assert_equals(bufferedEntries[1].duration, 0);
            assert_equals(bufferedEntries[2].entryType, "elementtiming");
            assert_equals(bufferedEntries[2].name, "element3");
            assert_equals(bufferedEntries[2].duration, 0);

            for (var name of ["element1", "element2", "element3"]) {
                const entriesByName = performance.getEntriesByName(name);
                assert_equals(entriesByName.length, 1, "Expected 1 elementtiming entry");
                assert_equals(entriesByName[0].entryType, "elementtiming");
                assert_equals(entriesByName[0].name, name);
                assert_equals(bufferedEntries[0].duration, 0);
            }
        }));
      }, "Basic test to check existence of text elementtiming entries");
    </script>
</body>
</html>
